requirements.txt
uvicorn
aiohttp
asusrouter
fastmcp